--- Devices/generic-esp32/Source/Configuration.cpp
@@ -1,3 +1,0 @@
-#include <Tactility/hal/Configuration.h>
-
-extern const tt::hal::Configuration hardwareConfiguration = {};
