+++ Devices/generic-esp32c6/device.properties
@@ -7,4 +7,6 @@ hardware.target=ESP32C6
 hardware.flashSize=8MB
 hardware.spiRam=false
 
+dependencies.useDeprecatedHal=false
+
 storage.userDataLocation=Internal
