+++ Devices/generic-esp32p4/CMakeLists.txt
@@ -1,7 +1,6 @@
-file(GLOB_RECURSE SOURCE_FILES Source/*.c*)
+file(GLOB_RECURSE SOURCE_FILES source/*.c*)
 
 idf_component_register(
     SRCS ${SOURCE_FILES}
-    INCLUDE_DIRS "Source"
-    REQUIRES Tactility
+    REQUIRES TactilityKernel driver
 )
