+++ Devices/generic-esp32p4/device.properties
@@ -7,4 +7,6 @@ hardware.target=ESP32P4
 hardware.flashSize=8MB
 hardware.spiRam=false
 
+dependencies.useDeprecatedHal=false
+
 storage.userDataLocation=Internal
