+++ Devices/generic-esp32p4/source/module.cpp
@@ -3,17 +3,15 @@
 extern "C" {
 
 static error_t start() {
-    // Empty for now
     return ERROR_NONE;
 }
 
 static error_t stop() {
-    // Empty for now
     return ERROR_NONE;
 }
 
-struct Module generic_esp32_module = {
-    .name = "generic-esp32",
+struct Module generic_esp32p4_module = {
+    .name = "generic-esp32p4",
     .start = start,
     .stop = stop,
     .symbols = nullptr,
