+++ Devices/generic-esp32s3/device.properties
@@ -7,4 +7,6 @@ hardware.target=ESP32S3
 hardware.flashSize=8MB
 hardware.spiRam=false
 
+dependencies.useDeprecatedHal=false
+
 storage.userDataLocation=Internal
